Added responsive css to chat-card when the sidenav is collapsed, moved the the message converted to the right parent directory and installed helmet npm package for html headers.....
