+++ api/package.json
@@ -17,6 +17,7 @@
     "dotenv": "^16.4.5",
     "express": "^4.21.0",
     "express-session": "^1.18.0",
+    "helmet": "^8.0.0",
     "langchain": "^0.3.2"
   },
   "devDependencies": {
