+++ src/app/utility/message-convert.ts
@@ -1,4 +1,4 @@
-import { Message } from "../../models/messages.models";
+import { Message } from "../models/messages.models";
 
 export function toMessage(role: 'system' | 'user' | 'assistant', content: string, system?: string) {
     return {
